requirements.txt
Python-jenkins
requests